% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21038R.R
\name{BayesianLasso}
\alias{BayesianLasso}
\title{A Bayesian Lasso Regression function with gibbs sampling}
\usage{
BayesianLasso(
  x,
  y,
  center = T,
  scale = T,
  a=1,
  b=1,
  n.max = 10000,
  E = TRUE,
  r = 1,
  d = 1
)
}
\arguments{
\item{x}{predictor variables}

\item{y}{response variable}

\item{center}{TRUE/FALSE (default: TRUE, if the design matrix x has been centered)}

\item{scale}{TRUE/FALSE (default: TRUE, if the design matrix x has been scaled)}
                         
\item{a, b}{parameter of sigma2's prior, default 1,1, if a=0,b=0, means the prior of sigma2 is 1/sigma2}

\item{n.max}{n of interations (default: 10000)}

\item{E}{TRUE/FALSE (default: TRUE, estimating lambda by empircal bayes; FALSE, estimating lambda by Hyperprior method.)}

\item{r, d}{hyper-Gamma prior for lambda^2 if E = FALSE}
}
\value{
beta: the median of matrix Beta.

tau2: the median of matrix Tau2

sigma2: the median of vector Sigma2

lambda: the median of vector Lambda
}
\description{
A function that can complete the Bayesian Lasso Regression and  can use empirical bayes and hyperpriors method.
}
\examples{
\dontrun{
data(Data)
attach(Data)
x<-Data$diabetes.x
y<-Data$diabetes.y
BayesianLasso(x,y,a=0,b=0)
}
}
\references{
Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.
}
