% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21038R.R
\name{Bayespredmse}
\alias{Bayespredmse}
\title{Prediction or MSE}
\usage{
Bayespredmse(
  x,
  y,
  x1 = as.matrix(0, 1, 1),
  center = T,
  scale = T,
  n.max = 10000,
  E = TRUE,
  r = 1,
  d = 1,
  pred = T
)
}
\arguments{
\item{x}{predictor variables}

\item{y}{response variable}

\item{x1}{a matrix which should be given if pred = T}

\item{center}{TRUE/FALSE (default: TRUE, if the design matrix x has been centered)}

\item{scale}{TRUE/FALSE (default: TRUE, if the design matrix x has been scaled)}

\item{n.max}{n of interations (default: 10000)}

\item{E}{TRUE/FALSE (default: TRUE, estimating lambda by empircal bayes; FALSE, estimating lambda by Hyperprior method.)}

\item{r, d}{hyper-Gamma prior for lambda^2 if E = FALSE}

\item{pred}{TRUE/FALSE (default: TRUE, we need the prediction; FALSE, we need the MSE)}
}
\value{
pred: the prediction

MSE: the MSE
}
\description{
A function that can complete the Bayesian Lasso Regression and get the prediction or MSE.
}
\examples{
\dontrun{
data(Data)
attach(Data)
x<-Data$diabetes.x
y<-Data$diabetes.y
Bayespredmse(x,y,pred=F)
}
}
\references{
Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.
}
