% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21038R.R
\name{Compare}
\alias{Compare}
\title{Compare with the existing function blasso in R package \code{monomvn}}
\description{
We compare the MSE which generates from the function we write with the function \code{blasso} in \code{monomvn}
}
\examples{
\dontrun{
library(monomvn)
data(Data)
attach(Data)
x<-Data$diabetes.x
y<-Data$diabetes.y
n<-nrow(Data$diabetes.x)
yc<-y-mean(y)
result<-BayesianLasso(x,y)
proc.time()
resid<-yc-x\%*\%result$beta
mse1<-t(resid)\%*\%resid/n
n.max=1e4
result1<-blasso(x,y,T=n.max)
proc.time()
result2<-round(apply(result1$beta[seq(round(n.max/2),n.max),],2, median),3)
resid<-yc-x\%*\%result2
mse2<-t(resid)\%*\%resid/n
mse1;mse2
}
}
