% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EmpiricalBayes}
\alias{EmpiricalBayes}
\title{Empirical Bayes method for lambda}
\usage{
EmpiricalBayes(L0, tol, tauinv2, p, lambda)
}
\arguments{
\item{L0}{a double value}

\item{tol}{a double value}

\item{tauinv2}{a vector}

\item{p}{an int value}

\item{lambda}{a double value}
}
\value{
a numericvector
}
\description{
it is an empirical bayes method to get the lambda, and we use mento carlo EM
}
\examples{
\dontrun{
L0<-0.1;tol<-10^(-3);tauinv2<-c(1,2,3);p<-3;lambda<-1
EmpiricalBayes(L0,tol,tauinv2,p,lambda)
}
}
