% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HyperpriorBayes}
\alias{HyperpriorBayes}
\title{Hyperprior method for lambda}
\usage{
HyperpriorBayes(tauinv2, r, d, p)
}
\arguments{
\item{tauinv2}{a vector}

\item{r}{an int value}

\item{d}{an int value}

\item{p}{an int vale}
}
\value{
a double value
}
\description{
Hyperprior method for lambda
}
\examples{
\dontrun{
tauinv2<-c(1,2,3);r<-1;d<-1;p<-3
HyperpriorBayes(tauinv2,r,d,p)
}
}
