\name{StatComp21038-package}
\alias{StatComp21038-package}
\alias{StatComp21038}
\docType{package}
\title{
  Bayesian Lasso Regression
}
\description{
  This package write a BayesianLasso function to solve the Bayesian Lasso Regression Problem by gibbs sampling.
}
\details{
  We use the function to solve the Bayesian Lasso Regression problem and compare it with the blasso function already existing in "monomvn"
}
\author{
Lifeng Liu

Maintainer: Lifeng Liu <cqllfustc@mail.ustc.edu.cn>
}
\references{
  Park, Trevor, and George Casella. "The bayesian lasso." Journal of the American Statistical Association 103.482 (2008): 681-686.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}

