% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumC}
\alias{sumC}
\title{A function that can sum the value of a vector}
\usage{
sumC(x, p)
}
\arguments{
\item{x}{the vector}

\item{p}{the length of x or the length we want to caculate}
}
\value{
a double value
}
\description{
A function that can sum the value of a vector
}
\examples{
\dontrun{
r<-as.numeric(-1,-2,3)
p<-length(r)
sumC(r,p)
}
}
